Add Dockerfile building ui and api images
