+++ apps/api/src/db/schemas/agreements.ts
@@ -19,9 +19,8 @@
 
 import { relations } from 'drizzle-orm';
 import { bigint, pgEnum, pgTable, text, timestamp, varchar } from 'drizzle-orm/pg-core';
-import { applications } from './applications.ts';
-// TODO: Integrate w/ TS
-// import { ApplicationAgreements } from 'pcgl-daco/packages/data-model/';
+
+import { applications } from './applications.js';
 
 export const agreementEnum = pgEnum('agreement_types', [
 	'dac_agreement_software_updates',
