+++ apps/api/src/db/schemas/files.ts
@@ -19,9 +19,8 @@
 
 import { relations } from 'drizzle-orm';
 import { bigint, customType, pgEnum, pgTable, timestamp, varchar } from 'drizzle-orm/pg-core';
-import { applications } from './applications.ts';
-// TODO: Integrate w/ TS
-// import { FileTypes } from 'pcgl-daco/packages/data-model/';
+
+import { applications } from './applications.js';
 
 export const fileTypesEnum = pgEnum('file_types', ['SIGNED_APPLICATION', 'ETHICS_LETTER']);
