+++ apps/api/src/main.ts
@@ -19,8 +19,8 @@
 
 import cors from 'cors';
 import express, { Request, Response } from 'express';
-// TODO: Fix Types package so we can import from main instead of specific file
-import { demoApplication } from 'pcgl-daco/packages/data-model/src/main.mts';
+
+import { demoApplication } from '@pcgl-daco/data-model';
 
 export const port = process.env.PORT || 3000;
