+++ apps/api/tsconfig.json
@@ -1,10 +1,10 @@
 {
-	"extends": "../../tsconfig.json",
+	"extends": "../tsconfig.apps.json",
 	"compilerOptions": {
 		"lib": ["ESNext"],
-		"module": "nodenext",
-		"moduleResolution": "nodenext",
-		"esModuleInterop": true
+		"target": "ESNext",
+		"module": "NodeNext",
+		"moduleResolution": "NodeNext"
 	},
 	"include": ["src", "./drizzle.config.ts"]
 }
