+++ apps/ui/docker/nginx.conf
@@ -1,0 +1,15 @@
+server {
+ listen 3000;
+ root   /usr/share/nginx/html;
+ index  index.html;
+ etag on;
+
+ location / {
+  try_files $uri $uri/ /index.html;
+ }
+
+ error_page   500 502 503 504  /50x.html;
+ location = /50x.html {
+  root   /usr/share/nginx/html;
+ }
+}
