+++ apps/ui/package.json
@@ -5,7 +5,7 @@
 	"type": "module",
 	"scripts": {
 		"dev": "vite",
-		"build": "tsc && vite build",
+		"build": "rimraf ./dist && tsc && vite build",
 		"lint": "eslint .",
 		"preview": "vite preview"
 	},
@@ -29,5 +29,9 @@
 		"typescript": "^5.5.3",
 		"typescript-eslint": "^8.0.1",
 		"vite": "^5.4.1"
+	},
+	"optionalDependencies": {
+		"@rollup/rollup-linux-x64-musl": "4.27.0",
+		"@rollup/rollup-linux-arm64-musl": "4.27.0"
 	}
 }
