+++ apps/ui/vite.config.ts
@@ -5,6 +5,10 @@ import { defineConfig } from 'vite';
 // https://vitejs.dev/config/
 export default defineConfig({
 	plugins: [react()],
+	build: {
+		outDir: './dist',
+		emptyOutDir: true,
+	},
 	resolve: {
 		alias: {
 			'@/assets': path.resolve(__dirname, './src/assets'),
