+++ package.json
@@ -2,7 +2,7 @@
 	"name": "pcgl-daco",
 	"version": "0.0.1",
 	"description": "Pan Canadian Genome Library DACO",
-	"main": "index.js",
+	"main": "dist/index.js",
 	"type": "module",
 	"homepage": "https://github.com/Pan-Canadian-Genome-Library/daco#readme",
 	"repository": {
@@ -13,6 +13,7 @@
 		"url": "https://github.com/Pan-Canadian-Genome-Library/daco/issues"
 	},
 	"scripts": {
+		"build:all": "pnpm -r build",
 		"test": "echo \"Error: no test specified\" && exit 1",
 		"prettier": "npx prettier ./packages ./apps --write",
 		"prepare": "husky"
@@ -24,6 +25,7 @@
 		"husky": "^9.1.6",
 		"prettier": "^3.3.3",
 		"prettier-plugin-organize-imports": "^4.1.0",
+		"rimraf": "^6.0.1",
 		"typescript": "^5.5.3",
 		"vite": "^5.4.1"
 	}
