+++ packages/data-model/package.json
@@ -2,16 +2,13 @@
 	"name": "@pcgl-daco/data-model",
 	"private": true,
 	"version": "0.0.0",
-	"main": "./src/main.mts",
+	"main": "./dist/main.js",
 	"type": "module",
-	"types": "./src/types.ts",
 	"scripts": {
 		"dev": "vite",
-		"build": "tsc && vite build",
-		"preview": "vite preview"
+		"build": "rimraf ./dist && tsc"
 	},
 	"devDependencies": {
-		"typescript": "^5.5.3",
-		"vite": "^5.4.1"
+		"typescript": "^5.5.3"
 	}
 }
