+++ tsconfig.base.json
@@ -1,0 +1,28 @@
+{
+	"compilerOptions": {
+		"target": "ESNext",
+		"module": "NodeNext",
+		"moduleResolution": "NodeNext",
+		"lib": ["ESNext"],
+		"moduleDetection": "force",
+		"skipLibCheck": true,
+		"esModuleInterop": true,
+		"allowSyntheticDefaultImports": true,
+		"forceConsistentCasingInFileNames": true,
+		"isolatedModules": true,
+		"useDefineForClassFields": true,
+		"outDir": "${configDir}/dist",
+		"declaration": true,
+		"sourceMap": true,
+		"declarationMap": true,
+
+		/* Linting */
+		"strict": true,
+		"noUnusedLocals": true,
+		"noImplicitAny": true,
+		"noUncheckedIndexedAccess": true,
+		"noFallthroughCasesInSwitch": true,
+
+		"noErrorTruncation": true
+	}
+}
